add tests for non-select updates
